Refactor: Move model class to repository package
