+++ index/api/routes.py
@@ -5,7 +5,7 @@ from fastapi import FastAPI, HTTPException
 from starlette.middleware.cors import CORSMiddleware
 from starlette.responses import RedirectResponse, HTMLResponse
 
-from index.db.model import Terminology, Concept, Mapping
+from index.repository.model import Terminology, Concept, Mapping
 from index.embedding import MPNetAdapter
 from index.repository.sqllite import SQLLiteRepository
 from index.visualisation import get_html_plot_for_current_database_state
