+++ index/process/ols.py
@@ -2,7 +2,7 @@ import logging
 
 import requests
 
-from index.db.model import Terminology, Concept, Mapping
+from index.repository.model import Terminology, Concept, Mapping
 from index.embedding import EmbeddingModel
 from index.repository.base import BaseRepository
