+++ tests/test_system.py
@@ -1,6 +1,6 @@
 import unittest
 
-from index.db.model import Terminology, Concept, Mapping
+from index.repository.model import Terminology, Concept, Mapping
 from index.embedding import MPNetAdapter
 from index.repository.sqllite import SQLLiteRepository
